requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.4.0
black==20.8b1
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
flake8==3.9.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
isort==5.7.0
kombu==5.0.2
mccabe==0.6.1
meilisearch==0.14.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
pdfminer.six==20200517
pdfplumber==0.5.27
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.0
pyparsing==2.4.7
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
Wand==0.6.6
wcwidth==0.2.5
whitenoise==5.2.0
